requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
idna==3.4
newsapi-python==0.2.7
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
